requirements.txt
redis
pycrypto
msgpack
netstruct